1.C.2.3 - Changed header file to match c file

+++ src/emulator/bitwiseshift.h
@@ -5,9 +5,9 @@
 #ifndef ARMV8_27_BITWISESHIFT_H
 #define ARMV8_27_BITWISESHIFT_H
 #include <stdint.h>
-uint32_t logicalLeftShift(uint32_t number, uint32_t shift);
-uint32_t logicalRightShift(uint32_t number, uint32_t shift);
-uint32_t arithmeticRightShift(uint32_t number, uint32_t shift);
-uint32_t rotateRight(uint32_t number, uint32_t shift);
+long long logicalLeftShift(long long number, uint32_t shift);
+long long logicalRightShift(long long number, uint32_t shift);
+long long arithmeticRightShift(long long number, uint32_t shift);
+long long rotateRight(long long number, uint32_t shift);
 
 #endif //ARMV8_27_BITWISESHIFT_H
